Renamed logging to logger to avoid circular import
